scripts/requirements.txt
# scripts/requirements.txt
anthropic>=0.79.0
python-dotenv==1.0.0
GitPython==3.1.40
